+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - added order number to manifest
 - added the possibility to change the manifest strings
 - added the possibility to change the manifest columns width
+- fixed error on PHP 8.2
 
 ## [1.0.10] - Fixes
 - removed PK service from "required offload postcode" services list